+++ Sstv.Domain.Sample/AssemblyInfo.cs
@@ -1,0 +1,9 @@
+using Sstv.Domain.Sample;
+using Sstv.DomainExceptions;
+
+[assembly: CollectErrorCodes(Types =
+[
+    typeof(ErrorCodes),
+    typeof(SecondErrorCodes),
+    typeof(DomainErrorCodes)
+])]